* MANIFEST add batchEditor.py and batchEditor.glade 	* glade/app.glade Add menu items for launching batch editor 	* glade/recCard.glade Fix alignment of items on Recipe Details page. 	* src/lib/GourmetRecipeManager.py Add batch editor. Create new 	methods to store models for attributes (cuisine, category, etc) in 	one place, rather than re-creating them in each recipe card, 	etc. that needs them. 	* src/lib/reccard.py Use new methods for creating global attribute 	models. 	* src/lib/recindex.py Fix type error that cropped up in update_recipe 	* src/lib/backends/sql_db.py 	* src/lib/importers/html_importer.py When dealing with pre-parsed 	ingredient dictionaries, only use our own parsing to fill in gaps 	(this allows html importer scripts to do some but not all parsing 	on their own).

+++ src/lib/recindex.py
@@ -615,7 +615,7 @@ class RecipeModel (pageable_store.PageableViewStore):
         for n,row in enumerate(self):
             debug('Looking at row',3)
             if row[0].id==recipe:
-                indx = n + (self.page * self.per_page)
+                indx = int(n + (self.page * self.per_page))
                 # update parent
                 self.parent_list[indx] = self.rd.fetch_one(self.rd.rview,
                                                            id=recipe)
